+++ Lommeregner/DoLommeregner.cs
@@ -25,14 +25,14 @@ namespace Lommeregner
             while (sentinel != "n")
             {
                 Console.WriteLine("Indtast tal 1: ");
-                int tal1 =  useri.ReadNumber();
+                double tal1 =  useri.ReadNumber();
                 
 
                 Console.WriteLine("Indtast en matematisk oprator[+ - * / %]: ");
                 string input2 = useri.ReadOperator();
 
                 Console.WriteLine("Indtast tal 2: ");
-                int tal2 = useri.ReadNumber();
+                double tal2 = useri.ReadNumber();
 
                 Console.WriteLine($"Resultat: " + calcu.mathcalc(tal1, tal2, input2));
 
@@ -45,8 +45,8 @@ namespace Lommeregner
           
 
 
-            //int tal2 = Int32.Parse(input2);
-            //bool erOperator = Int32.TryParse(input1, tal2);
+            //double tal2 = double32.Parse(input2);
+            //bool erOperator = double32.TryParse(input1, tal2);
 
         }
     }
